cleanup ideas:
TouchEngine could have a Background Layer input and only notify this Background if no touchitem is active. For that to work added ITouchItem.IsActive. The idea is that the TouchEngine can work out on Notify whether the notification has been processed. Got rid of FrameDelay crazyness in Example Manipulating shapes with gestures.
didn't try as I didn't have a touch device at hand
added a scetch on how to turn a layer into a touchItem based on its bounds.
added some comments